+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## Added
+- Use clustering of related objects to automatically create definitions
+
 ## Changed
 - Produce 2019 draft version schemas
 - Remove statistics from `NumberSchema` simple properties